IRUS V4 - Copy/requirements.txt
# Core requirements (required)
numpy
opencv-python
keyboard
mouse
pillow
scikit-learn
matplotlib
pandas

# Optional for advanced neural networks
# tensorflow